#78 공용 장바구니 식품 삭제 API 추가

+++ src/main/java/com/example/icebutler_server/cart/entity/multiCart/MultiCart.java
@@ -6,9 +6,6 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-import static javax.persistence.CascadeType.ALL;
 
 @NoArgsConstructor(access= AccessLevel.PROTECTED)
 @Getter
@@ -17,7 +14,7 @@ public class MultiCart extends BaseEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   @Column(nullable = false)
-  private int multiCartIdx;
+  private Long multiCartIdx;
 
   @OneToOne
   @JoinColumn(name = "multiFridgeUserIdx")
